servers/cluster03/files/home/opsuser/apps/mestech/requirements.txt
argparse==1.2.1
cassandra-driver==2.7.2
futures==3.0.3
kazoo==2.2.1
pykafka==2.0.2
pymssql==2.1.1
six==1.10.0
tabulate==0.7.5
thrift==0.9.3
wsgiref==0.1.2
